+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed bug where "r c" fails with disabled auto-min-io on a 4k block-sized SP
 - Clone: cleanup source snapshot if cloning failed
 - Clone: mark failed volumes with DELETE flags to allow to delete them afterward
+- Fixed --diskless-on-remaining not working with COMBINED nodes
 
 ## [1.33.0-rc.1] - 2025-11-11